+++ robottelo/cli/factory.py
@@ -68,8 +68,8 @@ from robottelo.cli.usergroup import UserGroupExternal
 from robottelo.cli.virt_who_config import VirtWhoConfig
 from robottelo.config import settings
 from robottelo.datafactory import valid_cron_expressions
-from robottelo.decorators import cacheable
 from robottelo.logging import logger
+from robottelo.utils.decorators import cacheable
 
 
 ORG_KEYS = ['organization', 'organization-id', 'organization-label']
