--- robottelo/decorators/func_shared/__init__.py
@@ -1,3 +1,0 @@
-from robottelo.decorators.func_shared.shared import shared  # noqa
-from robottelo.decorators.func_shared.shared import SharedFunctionError  # noqa
-from robottelo.decorators.func_shared.shared import SharedFunctionException  # noqa
