+++ robottelo/utils/decorators/func_shared/__init__.py
@@ -1,0 +1,3 @@
+from robottelo.utils.decorators.func_shared.shared import shared  # noqa
+from robottelo.utils.decorators.func_shared.shared import SharedFunctionError  # noqa
+from robottelo.utils.decorators.func_shared.shared import SharedFunctionException  # noqa
