+++ robottelo/utils/decorators/func_shared/base.py
